1. add getter and setter methods for user name and user email 2. connect to database to get check in data

+++ app/src/main/java/edu/northeastern/cs5520/numadfa21_happytravel/FriendAdpater.java
@@ -16,7 +16,7 @@ public class FriendAdpater extends RecyclerView.Adapter<FriendAdpater.FriendView
     private int[] images;
     private String[] names;
     private String[] places;
-    private int[] review_stars;
+    private float[] review_stars;
     private String[] review_contents;
 
     public FriendAdpater(
@@ -24,7 +24,7 @@ public class FriendAdpater extends RecyclerView.Adapter<FriendAdpater.FriendView
             int[] images,
             String[] names,
             String[] places,
-            int[] review_stars,
+            float[] review_stars,
             String[] review_contents) {
         this.context = ct;
         this.images = images;
